Add SARSA AutoTDP, profile sync fixes, GPD support, and UI improvements
- SARSA-based AutoTDP controller with learned game TDP store
- Fix AutoTDP state bleeding between per-game and global profiles
- Fix profile switch race conditions (debounce timer, slider/toggle echo)
- Fix UpdateTDPSliderEnabledState restoring stale LocalSettings values
- Use profile as source of truth for AutoTDP/TDPBoost toggle restore
- Add helper connection status indicator and quick metrics display
- Add GPD Win 5 controller and fan curve support
- Add RyzenSMU performance table and CO curve support
- Fix widget reconnection TDP mode reset
- Prevent LoadProfileSettings from overwriting helper values

+++ Samples/XboxGamingBar/Data/GPDButtonR4Property.cs
@@ -22,7 +22,7 @@ namespace XboxGamingBar.Data
         /// <param name="keycode">USB HID keycode.</param>
         public void SetKeycode(ushort keycode)
         {
-            SetValue(keycode);
+            SetValue((int)keycode);
         }
     }
 }
